moved hostID from OdeKafkaProperties to AppContext to keep scope of new class small and directed
