+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/kafka/OdeKafkaProperties.java
@@ -6,12 +6,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
-import us.dot.its.jpo.ode.eventlog.EventLogger;
 
-import java.net.InetAddress;
-import java.net.UnknownHostException;
 import java.util.Set;
-import java.util.UUID;
 
 @Configuration
 @ConfigurationProperties(prefix = "ode.kafka")
@@ -26,27 +22,4 @@ public class OdeKafkaProperties {
     private String producerType;
     @Value("${ode.kafka.disabled-topics:}")
     private Set<String> disabledTopics;
-
-    private String hostId;
-
-    public String getHostId() {
-        if (this.hostId == null || this.hostId.isEmpty()) {
-            initializeHostId();
-        }
-        return hostId;
-    }
-
-    private void initializeHostId() {
-        String hostname;
-        try {
-            hostname = InetAddress.getLocalHost().getHostName();
-        } catch (UnknownHostException e) {
-            // Let's just use a random hostname
-            hostname = UUID.randomUUID().toString();
-            logger.error("Unknown host error: {}, using random", e);
-        }
-        this.hostId = hostname;
-        logger.info("Host ID: {}", hostId);
-        EventLogger.logger.info("Initializing services on host {}", hostId);
-    }
 }
